+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   in the table (for [#1446](https://github.com/pbek/QOwnNotes/issues/1446))
 - fixed distraction free mode not restored properly on app restart if the app
   previously was maximized (for [#1448](https://github.com/pbek/QOwnNotes/issues/1448), thank you @Waqar144)
+- added more Indonesian, Italian, German, Russian translation (thank you zmni,
+  mlocati, tiwi90, rakekniven, unchase)
 
 ## 19.11.21
 - added a note list context menu item to jump to the note sub-folder of the